requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
requests==2.31.0
vk-api==11.9.9
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
typing_extensions==4.7.1
greenlet==2.0.2